+++ src/Login.ts
@@ -15,7 +15,7 @@ import { QuarkHandler, setPersonalQuarkBonus } from './Quark'
 import { updatePrestigeCount, updateReincarnationCount, updateTranscensionCount } from './Reset'
 import { format, player, saveSynergy } from './Synergism'
 import { Alert, Confirm, Notification } from './UpdateHTML'
-import { assert, btoa, isomorphicDecode, memoize } from './Utility'
+import { assert, btoa, displayHTMLError, isomorphicDecode, memoize } from './Utility'
 
 export type PseudoCoinConsumableNames = 'HAPPY_HOUR_BELL'
 
@@ -524,19 +524,7 @@ export async function handleLogin () {
                 if (response.redirected || response.ok) {
                   location.reload()
                 } else {
-                  const html = await response.text()
-                  const overlay = document.createElement('div')
-                  overlay.style.cssText =
-                    'position:fixed;inset:0;background:rgba(0,0,0,0.7);display:flex;align-items:center;justify-content:center;z-index:99999'
-                  const modal = document.createElement('div')
-                  modal.style.cssText =
-                    'background:var(--alert-color);color:var(--text-color);padding:20px;border:1px solid var(--boxmain-bordercolor);border-radius:8px;max-width:500px;max-height:80vh;overflow:auto'
-                  modal.innerHTML = DOMPurify.sanitize(html)
-                  overlay.addEventListener('click', (e) => {
-                    if (e.target === overlay) overlay.remove()
-                  })
-                  overlay.appendChild(modal)
-                  document.body.appendChild(overlay)
+                  await displayHTMLError(response)
                 }
               }
             } else {
@@ -855,6 +843,8 @@ export const renderCaptcha = platform === 'steam'
 
             if (response.redirected || response.ok) {
               location.reload()
+            } else {
+              await displayHTMLError(response)
             }
           } finally {
             form.dataset.submitting = undefined
